Refactor h5p iframe on master (#252)
* Add H5PIframe component
* Refactor H5PMedia
* Refactor Question
* Refactor Tapestry component
* Refactor LightBox
* Refactor EndScreen and QuizScreen
* Refactor h5p iframe and h5p media
