+++ templates/vue/src/components/Lightbox.vue
@@ -54,7 +54,6 @@
             @timeupdate="updateProgress"
             @complete="complete"
             @close="$emit('close')"
-            @h5p-recorder-saver-loaded="h5pRecorderSaverLoaded"
           />
         </div>
       </div>
@@ -83,10 +82,6 @@ export default {
   props: {
     nodeId: {
       type: [String, Number],
-      required: true,
-    },
-    tapestryApiClient: {
-      type: Object,
       required: true,
     },
   },
@@ -219,9 +214,6 @@ export default {
         ...this.dimensions,
         ...dimensions,
       }
-    },
-    h5pRecorderSaverLoaded(event) {
-      this.$emit("h5p-recorder-saver-loaded", { loadedH5pId: event.loadedH5pId })
     },
   },
 }
