+++ templates/vue/src/components/Tapestry.vue
@@ -36,10 +36,8 @@
     />
     <lightbox
       v-if="lightbox.isOpen"
-      :tapestry-api-client="TapestryAPI"
       :node-id="lightbox.id"
       @close="closeLightbox"
-      @h5p-recorder-saver-loaded="h5pRecorderSaverLoaded"
     />
   </div>
 </template>
@@ -63,7 +61,6 @@ export default {
   data() {
     return {
       loadedH5pId: 0,
-      recordedNodeIds: [],
       TapestryAPI: new TapestryApi(wpPostId),
       tapestryLoaded: false,
       modalType: "",
@@ -117,9 +114,6 @@ export default {
           return ["public", "authenticated"]
       }
     },
-    userLoggedIn: function() {
-      return wpApiSettings && wpApiSettings.userLoggedIn === "true"
-    },
     wpCanEditTapestry: function() {
       return wpApiSettings && wpApiSettings.wpCanEditTapestry === "1"
     },
@@ -130,7 +124,6 @@ export default {
     window.addEventListener("add-new-node", this.addNewNode)
     window.addEventListener("edit-node", this.editNode)
     window.addEventListener("tapestry-updated", this.tapestryUpdated)
-    window.addEventListener("tapestry-h5p-audio-recorder", this.saveH5PAudioToServer) // listen to event dispatched by H5P Audio Recorder lib
     window.addEventListener("open-lightbox", evt => this.openLightbox(evt.detail))
   },
   methods: {
@@ -144,62 +137,10 @@ export default {
       "updateNodeCoordinates",
     ]),
     ...mapActions(["addNode", "addLink", "updateNode", "updateNodePermissions"]),
-    async h5pRecorderSaverLoaded(event) {
-      this.loadedH5pId = event.loadedH5pId
-      const selectedNodeId = this.selectedNode.id
-      if (
-        selectedNodeId &&
-        this.loadedH5pId &&
-        this.recordedNodeIds.includes(selectedNodeId)
-      ) {
-        await this.loadH5PAudio(selectedNodeId, this.loadedH5pId)
-      }
-    },
-    async saveH5PAudioToServer(event) {
-      const encodedH5PAudio = event.detail.base64data.replace(
-        /^data:audio\/[a-z]+;base64,/,
-        ""
-      )
-      if (encodedH5PAudio && this.userLoggedIn) {
-        try {
-          const audio = {
-            blob: encodedH5PAudio,
-            h5pId: this.loadedH5pId,
-          }
-          await this.TapestryAPI.uploadAudioToServer(this.selectedNode.id, audio)
-          this.recordedNodeIds.push(this.selectedNode.id)
-        } catch (e) {
-          console.error(e)
-        }
-      }
-    },
-    async loadH5PAudio(nodeMetaId, loadedH5pId) {
-      try {
-        const audio = await this.TapestryAPI.getH5PAudioFromServer(
-          nodeMetaId,
-          loadedH5pId
-        )
-        const h5pAudioRecorder = document.getElementById("h5p")
-        if (h5pAudioRecorder) {
-          dispatchEvent(
-            new CustomEvent("tapestry-get-h5p-audio", {
-              detail: { audio },
-            })
-          )
-        } else {
-          console.error("H5P module is not loaded.")
-        }
-      } catch (e) {
-        console.error(e)
-      }
-    },
     async tapestryUpdated(event) {
       if (!this.tapestryLoaded) {
         this.init(event.detail.dataset)
         this.tapestryLoaded = true
-        if (this.userLoggedIn) {
-          this.recordedNodeIds = await this.TapestryAPI.getRecordedNodeIds()
-        }
       } else {
         this.setDataset(event.detail.dataset)
       }
