+++ templates/vue/src/components/lightbox/H5PMedia.vue
@@ -1,29 +1,23 @@
 <template>
   <div class="container">
-    <end-screen
-      :node="node"
-      :show="showEndScreen"
-      @rewatch="rewatch"
-      @close="close"
-      @h5p-recorder-saver-loaded="h5pRecorderSaverLoaded"
-    />
+    <end-screen :node="node" :show="showEndScreen" @rewatch="rewatch" @close="close" />
     <loading v-if="isLoading" label="Loading H5P media..." />
-    <iframe
-      id="h5p"
-      ref="h5p"
-      frameborder="0"
-      allowfullscreen="allowfullscreen"
-      :src="node.typeData.mediaURL"
+    <h5p-iframe
+      ref="h5pIframe"
+      :node="node"
       :width="width"
       :height="height"
-      @load="handleLoad"
-    ></iframe>
+      :settings="settings"
+      @is-loaded="isLoading = false"
+      @show-end-screen="showEndScreen = true"
+    />
   </div>
 </template>
 
 <script>
 import Loading from "../Loading"
 import EndScreen from "./EndScreen"
+import H5PIframe from "./H5PIframe"
 
 const ALLOW_SKIP_THRESHOLD = 0.95
 
@@ -31,6 +25,7 @@ export default {
   name: "h5p-media",
   components: {
     EndScreen,
+    'h5p-iframe': H5PIframe,
     Loading,
   },
   props: {
@@ -60,184 +55,11 @@ export default {
   methods: {
     rewatch() {
       this.showEndScreen = false
-      const h5pObj = this.$refs.h5p.contentWindow.H5P
-      const h5pVideo = h5pObj.instances[0].video
-      h5pVideo.seek(0)
-      h5pVideo.play()
+      this.$refs.h5pIframe.rewatch()
     },
     close() {
       this.showEndScreen = false
-      const h5pObj = this.$refs.h5p.contentWindow.H5P
-      const h5pVideo = h5pObj.instances[0].video
-      if (h5pVideo) {
-        h5pVideo.pause()
-      }
-      this.$emit("close")
-    },
-    handleLoad() {
-      this.isLoading = false
-
-      $("iframe").each(function() {
-        $(this)
-          .data("ratio", this.height / this.width)
-          // Remove the hardcoded width & height attributes
-          .removeAttr("width")
-          .removeAttr("height")
-      })
-      const setIframeDimensions = function() {
-        $("iframe").each(function() {
-          // Get the parent container's width
-          var width = $(this)
-            .parent()
-            .width()
-          var height = $(this)
-            .parent()
-            .height()
-          if (width * $(this).data("ratio") <= height) {
-            $(this)
-              .width(width)
-              .height(width * $(this).data("ratio"))
-          } else {
-            $(this)
-              .height(height)
-              .width(height / $(this).data("ratio"))
-          }
-        })
-      }
-      $(window).resize(setIframeDimensions)
-      setIframeDimensions()
-
-      const h5pObj = this.$refs.h5p.contentWindow.H5P
-      const mediaProgress = this.node.typeData.progress[0].value
-
-      // Check to see whether this is an H5P recorder
-      // If it is, we can emit an event to load the recorded audio (if exists)
-      // and terminate
-      if (h5pObj.instances[0].recorder) {
-        this.$emit("h5p-recorder-saver-loaded", {
-          loadedH5pId: h5pObj.instances[0].contentId,
-        })
-        return
-      }
-
-      if (this.node.mediaType === "video") {
-        const h5pVideo = h5pObj.instances[0].video
-        this.$emit("load", { el: h5pVideo })
-
-        const settings = this.settings
-
-        let seeked = false
-        let currentPlayedTime
-
-        h5pVideo.on("stateChange", event => {
-          switch (event.data) {
-            case h5pObj.Video.PLAYING: {
-              const videoDuration = h5pVideo.getDuration()
-              const updateVideoInterval = setInterval(() => {
-                if (
-                  currentPlayedTime !== h5pVideo.getCurrentTime() &&
-                  h5pVideo.getCurrentTime() > 0
-                ) {
-                  currentPlayedTime = h5pVideo.getCurrentTime()
-                  const amountViewed = currentPlayedTime / videoDuration
-
-                  this.$emit("timeupdate", "h5p", amountViewed)
-                  thisTapestryTool.updateProgressBars()
-
-                  if (amountViewed >= ALLOW_SKIP_THRESHOLD) {
-                    this.$emit("complete")
-                  }
-
-                  if (amountViewed >= 1) {
-                    this.showEndScreen = true
-                  }
-                } else {
-                  clearInterval(updateVideoInterval)
-                }
-              }, 300)
-
-              if (!seeked) {
-                // Change the video settings to whatever the user had set before
-                if (settings.volume !== undefined) {
-                  h5pVideo.setVolume(settings.volume)
-
-                  if (settings.muted) {
-                    h5pVideo.mute()
-                  } else {
-                    h5pVideo.unMute()
-                  }
-
-                  h5pVideo.setCaptionsTrack(settings.caption)
-                  h5pVideo.setQuality(settings.quality)
-                  h5pVideo.setPlaybackRate(settings.playbackRate)
-                }
-
-                const viewedAmount = mediaProgress * videoDuration
-                if (viewedAmount > 0) {
-                  h5pVideo.seek(viewedAmount)
-                }
-                if (viewedAmount === videoDuration) {
-                  this.showEndScreen = true
-                }
-                seeked = true
-              }
-
-              const { id, mediaType } = this.node
-              thisTapestryTool.updateMediaIcon(id, mediaType, "pause")
-              thisTapestryTool.recordAnalyticsEvent(
-                "user",
-                "play",
-                "h5p-video",
-                id,
-                { time: h5pVideo.getCurrentTime() }
-              )
-
-              break
-            }
-
-            case h5pObj.Video.PAUSED: {
-              const newSettings = {
-                volume: h5pVideo.getVolume(),
-                muted: h5pVideo.isMuted(),
-                caption: h5pVideo.getCaptionsTrack(),
-                quality: h5pVideo.getQuality(),
-                playbackRate: h5pVideo.getPlaybackRate(),
-                time: h5pVideo.getCurrentTime(),
-              }
-              seeked = true
-              const { id, mediaType } = this.node
-              thisTapestryTool.updateMediaIcon(id, mediaType, "play")
-              thisTapestryTool.recordAnalyticsEvent(
-                "user",
-                "pause",
-                "h5p-video",
-                id,
-                { time: h5pVideo.getCurrentTime() }
-              )
-              this.$emit("update-settings", newSettings)
-              break
-            }
-
-            case h5pObj.Video.BUFFERING: {
-              const { id, mediaType } = this.node
-              thisTapestryTool.updateMediaIcon(id, mediaType, "loading")
-              break
-            }
-          }
-        })
-        h5pVideo.play()
-        thisTapestryTool.recordAnalyticsEvent(
-          "app",
-          "auto-play",
-          "h5p-video",
-          this.node.id
-        )
-      }
-    },
-    h5pRecorderSaverLoaded(event) {
-      this.$emit("h5p-recorder-saver-loaded", {
-        loadedH5pId: event.loadedH5pId,
-      })
+      this.$refs.h5pIframe.close()
     },
   },
 }
